WAL-69: alter recipe_reviews.rating from integer to NUMERIC(3,1)
- EF migration 20260409130000_AlterRatingToNumeric: AlterColumn<decimal> with numeric(3,1); Down reverts to integer (half-star values would truncate on rollback — emergency use only)
- RecipeReview entity: Rating int → decimal
- ReviewDto + CreateReviewDto: Rating int → decimal
- RecipeDetailReviewDto: Rating int → decimal
- WalkerDbContextModelSnapshot updated to reflect new type

+++ backend/DTOs/RecipeDetailDto.cs
@@ -80,7 +80,7 @@ public class RecipeDetailReviewDto
     public int Id { get; init; }
     public int UserId { get; init; }
     public string UserName { get; init; } = string.Empty;
-    public int Rating { get; init; }
+    public decimal Rating { get; init; }
     public string? Notes { get; init; }
     public DateOnly? MadeOn { get; init; }
     public DateTime CreatedAt { get; init; }
